add day15.5 laravel

+++ database/migrations/2023_12_23_062207_create_cars_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('model');
             $table->boolean('active')->default(false);
-            $table->integer('price');
+            $table->double('price');
             $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->foreignId('prand_id')->references('id')->on('prands')->onDelete('cascade');
             $table->softDeletes();
